Rearrange timeout_transport code
